backend/services/chat_service/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# WebSocket support
python-socketio==5.11.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1

# HTTP client for AI service
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7
